# dictionary-server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary App for IT",
        "description": "API Docs for Dictionary",
        "contact": {
            "email": "quanghuybest@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/random-word": {
            "get": {
                "tags": [
                    "Words"
                ],
                "summary": "Lấy từ vựng ngẫu nhiên",
                "operationId": "ab71e39ada6e1fdef60730b7c2448934",
                "responses": {
                    "200": {
                        "description": "Lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm từ vựng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Hiện tại chưa có từ vựng!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/store-word": {
            "post": {
                "tags": [
                    "Words"
                ],
                "summary": "Thêm từ vựng mới",
                "description": "Thêm một từ mới khi người dùng nhập vào các input.",
                "operationId": "fb5a2dc96446512c21bb54cdf9380e58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "word_name",
                                    "specialization_id",
                                    "synonymous",
                                    "antonyms"
                                ],
                                "properties": {
                                    "word_name": {
                                        "description": "Tên từ vựng",
                                        "type": "string",
                                        "example": "Constant"
                                    },
                                    "specialization_id": {
                                        "description": "Id chuyên ngành.",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "synonymous": {
                                        "description": "Từ đồng nghĩa (có thể trống).",
                                        "type": "string",
                                        "example": "abc"
                                    },
                                    "antonyms": {
                                        "description": "Từ trái nghĩa (có thể trống).",
                                        "type": "string",
                                        "example": "xyz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thêm từ thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "word_name": {
                                            "type": "string",
                                            "example": "Constant"
                                        },
                                        "specialization_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "synonymous": {
                                            "type": "string",
                                            "example": "abc"
                                        },
                                        "antonyms": {
                                            "type": "string",
                                            "example": "xyz"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-27T04:58:55.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-10-27T04:58:55.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi trong quá trình xử lý request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Đã có lỗi xảy ra!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}